requirements.txt
# Azure AI Search RAG Toolkit - Dependencies

# Azure SDKs
azure-identity>=1.15.0
azure-search-documents>=11.6.0

# OpenAI
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=8.0.0
# pytest-cov>=4.1.0
# pytest-asyncio>=0.23.0
# ruff>=0.3.0
# mypy>=1.8.0
